requirements.txt
# Python dependencies for HFT project

# Data download
requests>=2.31.0

# Binance trading
python-binance>=1.0.34

# Data analysis (optional)
pandas>=2.0.0
numpy>=1.24.0
